backend/requirements.txt
Flask>=2.3.3
Flask-Cors>=4.0.0
Werkzeug>=2.3.7
PyJWT>=2.8.0
pandas>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
numpy>=1.24.0
